more reorganization of the cmd/registry/patch directory.
