+++ cmd/registry/patch/artifact.go
@@ -1,0 +1,95 @@
+// Copyright 2022 Google LLC. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package patch
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/apigee/registry/gapic"
+	"github.com/apigee/registry/log"
+	"github.com/apigee/registry/rpc"
+	"github.com/apigee/registry/server/registry/names"
+	"gopkg.in/yaml.v2"
+)
+
+type Artifact struct {
+	Header `yaml:",inline"`
+	Body   struct {
+		MimeType string `yaml:"mimeType,omitempty"`
+	} `yaml:"body"`
+}
+
+func buildArtifact(message *rpc.Artifact) (*Artifact, error) {
+	artifactName, err := names.ParseArtifact(message.Name)
+	if err != nil {
+		return nil, err
+	}
+	artifact := &Artifact{
+		Header: Header{
+			APIVersion: REGISTRY_V1,
+			Kind:       "Artifact",
+			Metadata: Metadata{
+				Name: artifactName.ArtifactID(),
+			},
+		},
+	}
+	return artifact, nil
+}
+
+// ExportArtifact writes an artifact as a YAML file.
+func ExportArtifact(ctx context.Context, client *gapic.RegistryClient, message *rpc.Artifact) {
+	bytes, _, err := exportArtifact(ctx, client, message)
+	if err != nil {
+		log.FromContext(ctx).WithError(err).Fatal("Failed to export artifact")
+	} else {
+		fmt.Println(string(bytes))
+	}
+}
+
+func exportArtifact(ctx context.Context, client *gapic.RegistryClient, message *rpc.Artifact) ([]byte, *Header, error) {
+	if message.Contents == nil {
+		req := &rpc.GetArtifactContentsRequest{
+			Name: message.Name,
+		}
+		body, err := client.GetArtifactContents(ctx, req)
+		if err != nil {
+			return nil, nil, err
+		}
+		message.Contents = body.Data
+	}
+	switch message.GetMimeType() {
+	case ManifestMimeType:
+		manifest, err := newManifest(message)
+		if err != nil {
+			return nil, nil, err
+		}
+		b, err := yaml.Marshal(manifest)
+		if err != nil {
+			return nil, nil, err
+		}
+		return b, &manifest.Header, nil
+	default:
+		artifact, err := buildArtifact(message)
+		if err != nil {
+			return nil, nil, err
+		}
+		b, err := yaml.Marshal(artifact)
+		if err != nil {
+			return nil, nil, err
+		}
+		return b, &artifact.Header, nil
+	}
+}
