+++ cmd/registry/patch/artifact_lifecycle.go
@@ -1,0 +1,30 @@
+// Copyright 2022 Google LLC. All Rights Reserved.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//    http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package patch
+
+type Lifecycle struct {
+	Header `yaml:",inline"`
+	Body   struct {
+		DisplayName string `yaml:"displayName,omitempty"`
+		Description string `yaml:"description,omitempty"`
+		Stages      []struct {
+			ID           string `yaml:"id"`
+			DisplayName  string `yaml:"displayName,omitempty"`
+			Description  string `yaml:"description,omitempty"`
+			URL          string `yaml:"url,omitempty"`
+			DisplayOrder int    `yaml:"displayOrder,omitempty"`
+		} `yaml:"stages"`
+	} `yaml:"body"`
+}
