+++ cmd/registry/patch/artifact_manifest.go
