+++ cmd/registry/patch/project.go
@@ -72,61 +72,3 @@ func ExportProject(ctx context.Context, client *gapic.RegistryClient, adminClien
 		}
 	})
 }
-
-// ExportAPI writes an API as a YAML file.
-func ExportAPI(ctx context.Context, client *gapic.RegistryClient, message *rpc.Api) {
-	api, err := buildAPI(ctx, client, message)
-	if err != nil {
-		log.FromContext(ctx).WithError(err).Fatal("Failed to export api")
-	}
-	b, err := yaml.Marshal(api)
-	if err != nil {
-		log.FromContext(ctx).WithError(err).Fatal("Failed to marshal doc as YAML")
-	}
-	fmt.Println(string(b))
-}
-
-// ExportArtifact writes an artifact as a YAML file.
-func ExportArtifact(ctx context.Context, client *gapic.RegistryClient, message *rpc.Artifact) {
-	bytes, _, err := exportArtifact(ctx, client, message)
-	if err != nil {
-		log.FromContext(ctx).WithError(err).Fatal("Failed to export artifact")
-	} else {
-		fmt.Println(string(bytes))
-	}
-}
-
-func exportArtifact(ctx context.Context, client *gapic.RegistryClient, message *rpc.Artifact) ([]byte, *Header, error) {
-	if message.Contents == nil {
-		req := &rpc.GetArtifactContentsRequest{
-			Name: message.Name,
-		}
-		body, err := client.GetArtifactContents(ctx, req)
-		if err != nil {
-			return nil, nil, err
-		}
-		message.Contents = body.Data
-	}
-	switch message.GetMimeType() {
-	case ManifestMimeType:
-		manifest, err := newManifest(message)
-		if err != nil {
-			return nil, nil, err
-		}
-		b, err := yaml.Marshal(manifest)
-		if err != nil {
-			return nil, nil, err
-		}
-		return b, &manifest.Header, nil
-	default:
-		artifact, err := buildArtifact(message)
-		if err != nil {
-			return nil, nil, err
-		}
-		b, err := yaml.Marshal(artifact)
-		if err != nil {
-			return nil, nil, err
-		}
-		return b, &artifact.Header, nil
-	}
-}
